# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/notifications": {
            "get": {
                "description": "Retrieves a paginated list of notifications for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get user notifications",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CapIot-api_internal_models.Notification"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
        "/api/notifications/delete-all": {
            "delete": {
                "description": "Deletes all notifications for the user",
                "tags": [
                    "notifications"
                ],
                "summary": "Delete all notifications",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
        "/api/notifications/device/{deviceID}": {
            "get": {
                "description": "Retrieves notifications for a specific device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get device notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "deviceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CapIot-api_internal_models.Notification"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
        "/api/notifications/mark-all-read": {
            "patch": {
                "description": "Marks all notifications for the user as read",
                "tags": [
                    "notifications"
                ],
                "summary": "Mark all notifications as read",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
        "/api/notifications/{notificationID}": {
            "delete": {
                "description": "Deletes a specific notification",
                "tags": [
                    "notifications"
                ],
                "summary": "Delete notification",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notification ID",
                        "name": "notificationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Marks a specific notification as read",
                "tags": [
                    "notifications"
                ],
                "summary": "Mark notification as read",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notification ID",
                        "name": "notificationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "CapIot-api_internal_models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Use the new ErrorCode type",
                    "allOf": [
                        {
                            "$ref": "#/definitions/CapIot-api_internal_models.ErrorCode"
                        }
                    ]
                },
                "details": {
                    "description": "Optional: Additional details"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                }
            }
        },
...
        "CapIot-api_internal_models.ErrorCode": {
            "type": "string",
            "enum": [
                "internal_server_error",
                "bad_request",
                "not_found",
                "forbidden",
                "unauthorized",
                "method_not_allowed",
                "invalid_token",
                "token_expired",
                "insufficient_permissions",
                "validation_failed",
                "missing_parameter",
                "invalid_format",
                "unsupported_media_type",
                "resource_not_found",
                "duplicate_resource"
            ],
            "x-enum-varnames": [
                "ErrorCodeInternalServerError",
                "ErrorCodeBadRequest",
                "ErrorCodeNotFound",
                "ErrorCodeForbidden",
                "ErrorCodeUnauthorized",
                "ErrorCodeMethodNotAllowed",
                "ErrorCodeInvalidToken",
                "ErrorCodeTokenExpired",
                "ErrorCodeInsufficientPermissions",
                "ErrorCodeValidationFailed",
                "ErrorCodeMissingParameter",
                "ErrorCodeInvalidFormat",
                "ErrorCodeUnsupportedMediaType",
                "ErrorCodeResourceNotFound",
                "ErrorCodeDuplicateResource"
            ]
        },
...
        "CapIot-api_internal_models.Notification": {
            "type": "object",
            "properties": {
                "component_name": {
                    "description": "Component name",
                    "type": "string"
                },
                "component_type": {
                    "description": "Component type (sensor, actuator, etc.)",
                    "type": "string"
                },
                "location_id": {
                    "description": "Associated location ID",
                    "type": "integer"
                },
                "location_name": {
                    "description": "Associated location name",
                    "type": "string"
                },
                "log_content": {
                    "description": "Content/message of the log",
                    "type": "string"
                },
                "log_id": {
                    "description": "Primary key in component_log table",
                    "type": "integer"
                },
                "log_read": {
                    "description": "Whether the notification has been read",
                    "type": "boolean"
                },
                "log_timestamp": {
                    "description": "Timestamp of the log",
                    "type": "string"
                },
                "log_type": {
                    "description": "Type of log (info, warning, error, etc.)",
                    "type": "string"
                },
                "site_id": {
                    "description": "Associated site ID",
                    "type": "integer"
                },
                "site_name": {
                    "description": "Associated site name",
                    "type": "string"
                }
            }
        },
...
    }
}